Create a messageApiId message attribute
