+++ src/components/content-tab-assistant.tsx
@@ -108,10 +108,10 @@ interface AuthenticatedViewProps {
 	isAssistantThinking: boolean;
 	updateMessage: OnUpdateMessageType;
 	siteId: string;
-	chatId: string;
 	handleSend: ( messageToSend?: string, isRetry?: boolean ) => void;
-	markMessageAsFeedbackReceived: ( id: number ) => void;
-	messageApiId: number;
+	markMessageAsFeedbackReceived: ReturnType<
+		typeof useAssistant
+	>[ 'markMessageAsFeedbackReceived' ];
 }
 
 export const AuthenticatedView = memo(
@@ -120,10 +120,8 @@ export const AuthenticatedView = memo(
 		isAssistantThinking,
 		updateMessage,
 		siteId,
-		chatId,
 		handleSend,
 		markMessageAsFeedbackReceived,
-		messageApiId,
 	}: AuthenticatedViewProps ) => {
 		const endOfMessagesRef = useRef< HTMLDivElement >( null );
 		const [ showThinking, setShowThinking ] = useState( false );
@@ -182,13 +180,11 @@ export const AuthenticatedView = memo(
 				siteId,
 				updateMessage,
 				message,
-				messageApiId,
 			}: {
 				message: MessageType;
 				showThinking: boolean;
 				siteId: string;
 				updateMessage: OnUpdateMessageType;
-				messageApiId: number;
 			} ) => {
 				const thinkingAnimation = {
 					initial: { opacity: 0, y: 20 },
@@ -235,12 +231,12 @@ export const AuthenticatedView = memo(
 											content={ message.content }
 										/>
 										<div className="flex justify-end">
-											{ !! message.id && ! message.feedbackReceived && (
+											{ !! message.id && !! message.messageApiId && ! message.feedbackReceived && (
 												<ChatRating
-													chatId={ chatId }
 													messageId={ message.id }
-													messageApiId={ messageApiId }
-													markMessageAsFeedbackReceived={ markMessageAsFeedbackReceived }
+													markMessageAsFeedbackReceived={ ( id, feedback ) =>
+														markMessageAsFeedbackReceived( id, feedback )
+													}
 													feedbackReceived={ !! message?.feedbackReceived }
 												/>
 											) }
@@ -253,7 +249,7 @@ export const AuthenticatedView = memo(
 					</>
 				);
 			},
-			[ chatId, markMessageAsFeedbackReceived ]
+			[ markMessageAsFeedbackReceived ]
 		);
 
 		if ( messages.length === 0 ) {
@@ -270,7 +266,6 @@ export const AuthenticatedView = memo(
 						updateMessage={ updateMessage }
 						message={ lastMessage }
 						showThinking={ showThinking }
-						messageApiId={ messageApiId }
 					/>
 				) }
 				<div ref={ endOfMessagesRef } />
@@ -337,7 +332,6 @@ export function ContentTabAssistant( { selectedSite }: ContentTabAssistantProps 
 	const { fetchAssistant, isLoading: isAssistantThinking } = useAssistantApi( selectedSite.id );
 	const { messages: welcomeMessages, examplePrompts } = useWelcomeMessages();
 	const [ input, setInput ] = useState< string >( '' );
-	const [ lastMessageApiId, setLastMessageApiId ] = useState< number | null >( null );
 	const isOffline = useOffline();
 	const { __ } = useI18n();
 	const lastMessage = messages.length === 0 ? undefined : messages[ messages.length - 1 ];
@@ -366,7 +360,7 @@ export function ContentTabAssistant( { selectedSite }: ContentTabAssistantProps 
 				const {
 					message,
 					chatId: fetchedChatId,
-					messageApiId: messageApiId,
+					messageApiId,
 				} = await fetchAssistant(
 					chatId,
 					[
@@ -376,8 +370,7 @@ export function ContentTabAssistant( { selectedSite }: ContentTabAssistantProps 
 					currentSiteChatContext
 				);
 				if ( message ) {
-					addMessage( message, 'assistant', chatId ?? fetchedChatId );
-					setLastMessageApiId( messageApiId );
+					addMessage( message, 'assistant', chatId ?? fetchedChatId, messageApiId );
 				}
 			} catch ( error ) {
 				if ( typeof messageId !== 'undefined' ) {
@@ -433,16 +426,16 @@ export function ContentTabAssistant( { selectedSite }: ContentTabAssistantProps 
 								examplePrompts={ examplePrompts }
 								disabled={ disabled }
 							/>
-							<AuthenticatedView
-								messages={ messages }
-								isAssistantThinking={ isAssistantThinking }
-								updateMessage={ updateMessage }
-								markMessageAsFeedbackReceived={ markMessageAsFeedbackReceived }
-								siteId={ selectedSite.id }
-								chatId={ chatId }
-								handleSend={ handleSend }
-								messageApiId={ lastMessageApiId }
-							/>
+							{
+								<AuthenticatedView
+									messages={ messages }
+									isAssistantThinking={ isAssistantThinking }
+									updateMessage={ updateMessage }
+									markMessageAsFeedbackReceived={ markMessageAsFeedbackReceived }
+									siteId={ selectedSite.id }
+									handleSend={ handleSend }
+								/>
+							}
 						</>
 					) : (
 						! isOffline && <UnauthenticatedView onAuthenticate={ authenticate } />
